snowflake-project/requirements.txt
snowflake-connector-python
sqlparse
pytest